requirements.txt
pandas==1.5.3
openai==0.27.10
gspread==5.10.0
python-dotenv==1.0.0
requests==2.31.0
-e .